Initial commit.
This code is a COMPLETE MESS. The disassembler will MAKE YOUR EYES
BLEED!

So far it's just a simple in-place substitution engine, most .exes don't
provide relocations, and it would be completely unsafe to attempt an
incomplete relocation. Atm we only have limited static analysis, we have
no way to know where CALL EAX is going to take us, even if the address
is a plaintext constant moved into EAX two bytes ago.
We assume no self-modifying code, and attempt to follow the code flow
throught the branching instructions.

The analysis atm is just builds a list of all branching instructions
(regardless of the fact that we don't know where some of them lead), a
multimap of known xref-ed addresses, and an IDA-like jump flow graph
(doesn't take calls/ints into account, every block is "jump-atomic").

If your program is packed but not protected by fancy obfuscations such
as having the EP directly trigger a SEH or bypassing R^X and jumping
outside of the .text, you're in trouble. Hopefully the disassembler will
land on something it can't handle and throw an exception. You can use -f
to continue anyway, but you probably don't want to.

Also right now it can't handle anything that doesn't have its entry
point in the .text section. Having a CODE section will fail, and an
.itext section merged with the .text section will fail horribly.

The only transform is dirt-simple inplace substitution of some
instructions by equivalent instructions. E.g swapping 0x80 with 0x82,
randomizing the scale of the SIB when the index is EBP (or removing the
SIB), swapping src/dst registers and using an instruction that reads
src/dst backwards, etc.

Somehow, it works.