Disable analysis, add simple data detection
Add a simple detection of data to the disassembler : An address
referenced in a MOV Zv Iv instruction, but not referenced by a branching
instruction will be marked as possible data, and we won't attempt to
decompile it.

Also disables the separate analysis (branches list, jump flow graph)
since the simple inplace substitution isn't using them.

We still can't detect noreturn functions, but if there's referenced data
after the noreturn call, we might still detect it and stop.