Add option to encrypt a section
Add option "-e sectionName" that will encrypt a section (currently a
simple xor, since it shouldn't be too suspicious. Everyone uses xor for
everything).

It picks a decryptor template (currently only 2 available : one not
obfuscated and not too suspicious, and one that obfuscates the original
entry point a little), and fills the gaps.

The key is always random, and some decryptor templates can have other
random elements as well, so the resulting code never has the same
checksum.

Note that this is only intended to block simple static analysis, not to
be a FUD polymorphic crypter, any AV for example could just sandbox us
until we finish decrypting ourselves.

The purpose it to have a basic checksum-breaking algorythm on the data,
since we can already use substitutions on the code.

The decryptor is in its own section, and flagged as the new entry point.

Almost all the pointers to the virtual image were replaced by offsets.
We can now add sections to the image, but this requires a realloc of the
data and virtual image.